+++ AutomatedLabTest/internal/tests/CaSubordinate.tests.ps1
@@ -1,15 +1,8 @@
-﻿param
-(
-    [Parameter(Mandatory)]
-    [AutomatedLab.Lab]
-    $Lab
-)
-
-Describe "$($Lab.Name) CaSubordinate" -Tag CaSubordinate {
-
-    Context "Role deployment successful" {
-        It "Should return the correct amount of machines" {
-            (Get-LabVm -Role CaSubordinate).Count | Should -Be $Lab.Machines.Where({$_.Roles.Name -contains 'CaSubordinate'}).Count
+Describe "[$($(Get-Lab).Name)] CaSubordinate" -Tag CaSubordinate {
+Context "Role deployment successful" {
+        It "[CaSubordinate] Should return the correct amount of machines" {
+            (Get-LabVm -Role CaSubordinate).Count | Should -Be $(Get-Lab).Machines.Where({$_.Roles.Name -contains 'CaSubordinate'}).Count
         }
     }
 }
+