+++ AutomatedLabTest/internal/tests/DC.tests.ps1
@@ -1,15 +1,7 @@
-﻿param
-(
-    [Parameter(Mandatory)]
-    [AutomatedLab.Lab]
-    $Lab
-)
-
-Describe "$($Lab.Name) DC" -Tag DC {
-
-    Context "Role deployment successful" {
-        It "Should return the correct amount of machines" {
-            (Get-LabVm -Role DC).Count | Should -Be $Lab.Machines.Where({$_.Roles.Name -contains 'DC'}).Count
+Describe "[$($(Get-Lab).Name)] DC" -Tag DC {
+Context "Role deployment successful" {
+        It "[DC] Should return the correct amount of machines" {
+            (Get-LabVm -Role DC).Count | Should -Be $(Get-Lab).Machines.Where({$_.Roles.Name -contains 'DC'}).Count
         }
     }
 }