+++ AutomatedLabTest/internal/tests/FileServer.tests.ps1
@@ -1,15 +1,8 @@
-﻿param
-(
-    [Parameter(Mandatory)]
-    [AutomatedLab.Lab]
-    $Lab
-)
-
-Describe "$($Lab.Name) FileServer" -Tag FileServer {
-
-    Context "Role deployment successful" {
-        It "Should return the correct amount of machines" {
-            (Get-LabVm -Role FileServer).Count | Should -Be $Lab.Machines.Where({$_.Roles.Name -contains 'FileServer'}).Count
+Describe "[$($(Get-Lab).Name)] FileServer" -Tag FileServer {
+Context "Role deployment successful" {
+        It "[FileServer] Should return the correct amount of machines" {
+            (Get-LabVm -Role FileServer).Count | Should -Be $(Get-Lab).Machines.Where({$_.Roles.Name -contains 'FileServer'}).Count
         }
     }
 }
+